Set log location at '/var/log/multibootusb.log' and show that info before installation of a distro commences. Use 'RotatingFileHandler' instead of 'os.remove()' to rotate log because trying to remove file without closing the open handle causes exception on Windows. Move a handleful of funtions from gen.py to osdriver.py to avoid cyclic dependency.
